+++ pkg/schedule/core/cluster_informer.go
@@ -17,9 +17,9 @@ package core
 import (
 	"github.com/tikv/pd/pkg/core"
 	sc "github.com/tikv/pd/pkg/schedule/config"
+	"github.com/tikv/pd/pkg/schedule/keyrange"
 	"github.com/tikv/pd/pkg/schedule/labeler"
 	"github.com/tikv/pd/pkg/schedule/placement"
-	"github.com/tikv/pd/pkg/schedule/rangelist"
 	"github.com/tikv/pd/pkg/statistics"
 	"github.com/tikv/pd/pkg/statistics/buckets"
 	"github.com/tikv/pd/pkg/storage"
@@ -62,7 +62,7 @@ type SharedCluster interface {
 	GetBasicCluster() *core.BasicCluster
 	GetSharedConfig() sc.SharedConfigProvider
 	GetRuleManager() *placement.RuleManager
-	GetKeyRangeManager() *rangelist.KeyRangeManager
+	GetKeyRangeManager() *keyrange.KeyRangeManager
 	AllocID(uint32) (uint64, uint32, error)
 	IsSchedulingHalted() bool
 }
