+++ pkg/schedule/keyrange/key_range_manager.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package rangelist
+package keyrange
 
 import (
 	"sync"
