+++ server/cluster/cluster.go
@@ -19,6 +19,7 @@ import (
 	"encoding/json"
 	errorspkg "errors"
 	"fmt"
+	"github.com/tikv/pd/pkg/schedule/keyrange"
 	"io"
 	"math"
 	"net/http"
@@ -59,7 +60,6 @@ import (
 	"github.com/tikv/pd/pkg/schedule/hbstream"
 	"github.com/tikv/pd/pkg/schedule/labeler"
 	"github.com/tikv/pd/pkg/schedule/placement"
-	"github.com/tikv/pd/pkg/schedule/rangelist"
 	"github.com/tikv/pd/pkg/slice"
 	"github.com/tikv/pd/pkg/statistics"
 	"github.com/tikv/pd/pkg/statistics/utils"
@@ -176,7 +176,7 @@ type RaftCluster struct {
 	opt *config.PersistOptions
 	*schedulingController
 	ruleManager              *placement.RuleManager
-	keyRangeManager          *rangelist.KeyRangeManager
+	keyRangeManager          *keyrange.KeyRangeManager
 	regionLabeler            *labeler.RegionLabeler
 	replicationMode          *replication.ModeManager
 	unsafeRecoveryController *unsaferecovery.Controller
@@ -323,7 +323,7 @@ func (c *RaftCluster) InitCluster(
 	c.keyspaceGroupManager = keyspaceGroupManager
 	c.hbstreams = hbstreams
 	c.ruleManager = placement.NewRuleManager(c.ctx, c.storage, c, c.GetOpts())
-	c.keyRangeManager = rangelist.NewKeyRangeManager()
+	c.keyRangeManager = keyrange.NewKeyRangeManager()
 	if c.opt.IsPlacementRulesEnabled() {
 		err := c.ruleManager.Initialize(c.opt.GetMaxReplicas(), c.opt.GetLocationLabels(), c.opt.GetIsolationLevel(), false)
 		if err != nil {
@@ -953,7 +953,7 @@ func (c *RaftCluster) GetRuleManager() *placement.RuleManager {
 }
 
 // GetKeyRangeManager returns the key range manager reference
-func (c *RaftCluster) GetKeyRangeManager() *rangelist.KeyRangeManager {
+func (c *RaftCluster) GetKeyRangeManager() *keyrange.KeyRangeManager {
 	return c.keyRangeManager
 }
